requirements.txt
webuntis
openai
python-telegram-bot
python-dotenv
